requirements.txt
Django==1.10.2
djangorestframework==3.5.1
Pillow==3.4.2
